Updated edit functions
Issue: After editing an event, the event descriptions on IPhones would show up as HTML
Looking into it: whenever the date or time was edited, the description and location are grabbed from the event before creating a new event. This data is transferred to the new event as an HTML string. Apple calendars does not support HTML strings so it just shows up as raw HTML.
Solution: Need to reconstruct the event location and description as text string every time there is an edit. This means updating each of the 3 unctions that handle edits and adding a function to build the new descriptions based on current sheet data